Add experimental nix-darwin & home-manager config
I've been using this for some time, but haven't captured it in git. I
thought it was about time I did. Unfortunately this means I've lost the
history behind this file and consequently the reasoning behind it all!

The idea is that this replaces some of the contents of this dotfiles
repo, but I haven't yet got round to deleting the bits that I've moved
across. It also captures some system configuration that I previously
recorded in this Gist [1].

I'm still using homebrew to manage a lot of packages and asdf to manage
version of packages like Ruby.

[1]: https://gist.github.com/floehopper/80a604e5ca750641725a56bb01e8b409