Move vimrc into home-manager config
Although I hadn't run the script to set up the sym link to vimrc, I
think these settings are probably useful.

Note that the `extraConfig` is stored in a vimrc file in the nix store
rather than in `~/.vimrc`. You can see the path to this if you run the
`:scriptnames` command in vim.