Remove unused git config files
These git config files were moved into into the home-manager config some
time ago and I hadn't run the script to set up the sym link to them on
this machine. So it's safe to delete them. There might be a few that I
didn't copy over, but I've been doing without them for some time!

Note that these files can be found in the `~/.config/git/` directory.
The `~/.config` path is the default under the XDG Directory
Specification.