Move IRB config into home-manager
Most of the config that was in `~/.irbrc` is now the default in recent
versions of Ruby. The only thing that isn't is disabling auto-complete.
I think I had disabled this, because I found it very slow when SSH-ed
into in a GOV.UK with a large codebase. So I've commented it out for now
so I can try having it enabled again, but easily disable it if it
becomes annoying.